requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosend==3.0.5
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.1.0
certifi==2024.12.14
charset-normalizer==3.4.5
Deprecated==1.3.1
distro==1.9.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
izulu==0.50.0
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
netaddr==1.3.0
packaging==26.0
propcache==0.2.1
py3xui==0.5.5
pycron==3.2.0
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==7.3.0
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.37
taskiq-dependencies==1.5.7
typing-inspection==0.4.1
typing_extensions==4.12.2
tzlocal==5.3.1
urllib3==2.6.3
wrapt==2.1.2
yarl==1.18.3
yookassa==3.10.0
